--- packages/components/src/DatePicker/DayPicker/DateDayPickerGrid/index.ts
@@ -1,1 +1,0 @@
-export * from './styles';
