--- packages/components/src/DatePicker/DayPicker/DateDayPickerGridHead/index.ts
@@ -1,1 +1,0 @@
-export * from './DateDayPickerGridHead';
