+++ packages/components/src/DatePicker/DayPicker/DayPicker.tsx
@@ -12,13 +12,13 @@ import { type PickerProps } from '../types';
 import { DateCompareDeep, addMonths } from '../../utils/date';
 import { useLocaleDateTimeFormat } from '../hooks/useLocaleDateTimeFormat';
 import { ConfigContext } from '../../ConfigProvider';
-import { DAYS_IN_WEEK } from '../constants/counts';
+import { DAYS_IN_WEEK } from '../constants';
 import { isDateBetweenSelectedAndRangeDates } from '../utils';
 import { PopoverHoveredContext } from '../PopoverHoveredContext';
 
-import { DateDayPickerGridHead } from './DateDayPickerGridHead';
-import { DateDayPickerGridBody } from './DateDayPickerGrid';
-import { useDaysGrid } from './hooks/useDaysGrid';
+import { Head } from './Head';
+import { Body } from './styles';
+import { useDaysGrid } from './hooks';
 
 export type MondayFirst = {
   /**
@@ -97,8 +97,8 @@ export const DayPicker = ({
         headBtnText={monthYearFormat(baseDate)}
       />
       <DateCalendarBody>
-        <DateDayPickerGridHead isMondayFirst={isMondayFirst} />
-        <DateDayPickerGridBody role="grid">
+        <Head isMondayFirst={isMondayFirst} />
+        <Body role="grid">
           {grid.map(({ date, monthDay, ...props }, index) => (
             <DateCalendarGridButton
               key={index}
@@ -124,7 +124,7 @@ export const DayPicker = ({
               {monthDay}
             </DateCalendarGridButton>
           ))}
-        </DateDayPickerGridBody>
+        </Body>
       </DateCalendarBody>
     </DateCalendarWrapper>
   );
