+++ packages/components/src/DatePicker/DayPicker/Head/index.ts
@@ -1,0 +1,1 @@
+export * from './Head';
