+++ packages/components/src/DatePicker/DayPicker/Head/styles.ts
@@ -1,5 +1,5 @@
 import { styled } from '../../../styles';
-import { DAYS_IN_WEEK } from '../../constants/counts';
+import { DAYS_IN_WEEK } from '../../constants';
 
 export const Wrapper = styled.div`
   display: grid;
