+++ packages/components/src/DatePicker/DayPicker/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useDaysGrid';
