+++ packages/components/src/DatePicker/MinMaxDateContext/MinMaxDateContextProvider/index.ts
@@ -1,3 +1,1 @@
-export * from './MinMaxDateContext';
-
 export * from './MinMaxDateContextProvider';
