+++ packages/components/src/DatePicker/YearPicker/YearPicker.tsx
@@ -11,9 +11,9 @@ import { useCalendarNavigate } from '../hooks/useCalendarNavigate';
 import { addYears } from '../../utils/date';
 import { type PickerProps } from '../types';
 import { ConfigContext } from '../../ConfigProvider';
-import { ELEMENTS_COUNT_IN_ROW_IN_LARGE_GRID } from '../constants/counts';
+import { ELEMENTS_COUNT_IN_ROW_IN_LARGE_GRID } from '../constants';
 
-import { useYearsGrid } from './hooks/useYearsGrid';
+import { useYearsGrid } from './hooks';
 import { YEARS_IN_GRID } from './constants';
 
 type DateYearPickerProps = PickerProps & CommonDateCalendarHeadProps;
